refactor: rename and move IcReceiveButton
