+++ src/frontend/src/icp/components/receive/IcReceiveICP.svelte
@@ -4,7 +4,7 @@
 	import { isRouteTokens } from '$lib/utils/nav.utils';
 	import { page } from '$app/stores';
 	import IcReceiveInfoICP from '$icp/components/receive/IcReceiveInfoICP.svelte';
-	import IcReceiveButton from '$icp/components/receive/IcReceiveButton.svelte';
+	import ReceiveButton from '$lib/components/receive/ReceiveButton.svelte';
 	import { modalIcpReceive } from '$lib/derived/modal.derived';
 	import { getContext } from 'svelte';
 	import {
@@ -24,7 +24,7 @@
 	};
 </script>
 
-<IcReceiveButton on:click={openReceive} />
+<ReceiveButton on:click={openReceive} />
 
 {#if $modalIcpReceive}
 	<IcReceiveModal infoCmp={IcReceiveInfoICP} />
